fix: categorized landing page data into their own folders, refactored Banner.tsx, and polished up landing page
