+++ lib/components/landing-page/Footer.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { useRouter } from 'next/router';
 
 /**
