+++ lib/components/landing-page/Header.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import { useState } from 'react';
+import React, { useState } from 'react';
 
 const Links: React.FC = () => {
   const router = useRouter();
