+++ lib/components/landing-page/assets/data/promoData.ts
