+++ lib/components/landing-page/assets/img/bg.png
