+++ lib/components/landing-page/index.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import Banner from './Banner';
 import Footer from './Footer';
 import Header from './Header';
