+++ lib/components/landing-page/team-page/Card.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import Links from './Links';
 
 interface TeamMemberType {
