+++ lib/components/landing-page/team-page/profileMobile.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import 'react-responsive-carousel/lib/styles/carousel.min.css'; // requires a loader
 
 function ProfileMobile(props) {
